ref: Split options into modules and use @sentry/react in the browser
