+++ packages/nextjs/src/sdk.ts
@@ -1,6 +1,7 @@
-import { init as browserInit } from '@sentry/browser';
+import { init as reactInit } from '@sentry/react';
 
-import { MetadataBuilder, NextjsOptions } from './options';
+import { MetadataBuilder } from './utils/metadataBuilder';
+import { NextjsOptions } from './utils/nextjsOptions';
 
 /**
  * The Sentry NextJS SDK Client.
@@ -11,5 +12,5 @@ import { MetadataBuilder, NextjsOptions } from './options';
 export function init(options: NextjsOptions): void {
   const metadataBuilder = new MetadataBuilder(options, ['nextjs']);
   metadataBuilder.addSdkMetadata();
-  browserInit(options);
+  reactInit(options);
 }
