+++ packages/nextjs/src/utils/nextjsOptions.ts
@@ -1,0 +1,7 @@
+import { NodeOptions } from '@sentry/node';
+import { BrowserOptions } from '@sentry/react';
+import { Options } from '@sentry/types';
+
+export interface NextjsOptions extends Options, BrowserOptions, NodeOptions {
+  // TODO: options for NextJS
+}
